rename VeProtocol to ve_protocol
